manipmongo: fix pagination
The pagination management with next and ordering was a mess and full or
bugs anfd gotchas.
This patch moves the retrieve many implementation to a aggregate
pipeline, where it is actually possible to correct order, then match.

This also optimizes how we do sharding, and namespace filtering by
running their match before doing anything else.

The code has been realigned and refactored, and the refactored functions
are now tested.