manipcli: fix panic in setViperFlagsWithPrefix
In setViperFlagsWithPrefix() we sitch on the spec type and we assume that if the type is ref that the instance type must be a pointer to a struct. However, in the case of the type being just a struct, the reflect call to Elem() will panic as the kind is not a pointer.